# 03_backend/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Restausync API
  description: Sistema integral para gestión de restaurantes (usuarios, pedidos, cocina, inventario y clientes)
  version: 1.0.0
  contact:
    name: JEBC-DeV
    email: jebcdev@example.com
servers:
  - url: http://localhost:4003/api/v1
    description: Servidor de desarrollo
  - url: https://api.restaurant-sync.com/api/v1
    description: Servidor de producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Error description"
        data:
          type: object
          description: "Additional error data"
...
    Rol:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          enum: [admin, chef, mesero, inventario]
          example: "mesero"
        description:
          type: string
          example: "Usuario mesero"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@admin.com"
        password:
          type: string
          example: "12345678"
    LoginResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/Rol'
        created_at:
          type: string
          format: date-time
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
paths:
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión de usuario
      description: Autentica a un usuario y devuelve un token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...